Feedback hw02
Good work! See my comments marked with BMW

In many places, your code produced results that should have been definitely not what you wanted. be sure that you carefully look at the output it gives you and that it is correct. It's not enough that the code runs, it needs to also give you the correct answer.

Also, you only did have of the question for Exercises 2.1 and 2.2 Please be sure to ready the directions carefully.